Rename container_docker_common.h to container_docker.h, fix/cleanup includes
