+++ userspace/libsinsp/container_docker.h
