+++ userspace/libsinsp/container_docker_win.cpp
@@ -17,6 +17,12 @@ limitations under the License.
 
 */
 
+#include "container_docker.h"
+#include "sinsp.h"
+#include "sinsp_int.h"
+#include "dragent_win_hal_public.h"
+
+
 sinsp_container_engine_docker::sinsp_container_engine_docker() :
 	m_api_version("/v1.30")
 {
